CLIP/clip_requirements.txt
torch==2.7.1
torchvision==0.22.1
open_clip_torch
Pillow
tqdm
scikit-learn
numpy 